+++ tests/expand/tests/expand/naming-enum-ref.expanded.rs
@@ -17,6 +17,7 @@ enum Enum<T, U> {
 #[allow(clippy::pattern_type_mismatch)]
 #[allow(clippy::redundant_pub_crate)]
 #[allow(dead_code)]
+#[allow(clippy::ref_option_ref)]
 #[allow(clippy::type_repetition_in_bounds)]
 enum ProjRef<'pin, T, U>
 where