+++ tests/expand/tests/expand/naming-struct-own.expanded.rs
@@ -51,6 +51,7 @@ const _: () = {
     #[allow(clippy::pattern_type_mismatch)]
     #[allow(clippy::redundant_pub_crate)]
     #[allow(dead_code)]
+    #[allow(clippy::ref_option_ref)]
     #[allow(clippy::type_repetition_in_bounds)]
     struct __StructProjectionRef<'pin, T, U>
     where