+++ tests/expand/tests/expand/naming-tuple_struct-all.expanded.rs
@@ -22,6 +22,7 @@ where
 #[allow(clippy::pattern_type_mismatch)]
 #[allow(clippy::redundant_pub_crate)]
 #[allow(dead_code)]
+#[allow(clippy::ref_option_ref)]
 #[allow(clippy::type_repetition_in_bounds)]
 struct ProjRef<'pin, T, U>(::pin_project::__private::Pin<&'pin (T)>, &'pin (U))
 where