+++ tests/expand/naming/struct-own.expanded.rs
@@ -6,6 +6,7 @@ struct Struct<T, U> {
     unpinned: U,
 }
 #[allow(box_pointers)]
+#[allow(deprecated)]
 #[allow(explicit_outlives_requirements)]
 #[allow(single_use_lifetimes)]
 #[allow(unreachable_pub)]
@@ -18,6 +19,7 @@ struct ProjOwn<T, U> {
     unpinned: U,
 }
 #[allow(box_pointers)]
+#[allow(deprecated)]
 #[allow(explicit_outlives_requirements)]
 #[allow(single_use_lifetimes)]
 #[allow(unreachable_pub)]
@@ -27,6 +29,7 @@ struct ProjOwn<T, U> {
 #[allow(clippy::used_underscore_binding)]
 const _: () = {
     #[allow(box_pointers)]
+    #[allow(deprecated)]
     #[allow(explicit_outlives_requirements)]
     #[allow(single_use_lifetimes)]
     #[allow(unreachable_pub)]
@@ -44,6 +47,7 @@ const _: () = {
         unpinned: &'pin mut (U),
     }
     #[allow(box_pointers)]
+    #[allow(deprecated)]
     #[allow(explicit_outlives_requirements)]
     #[allow(single_use_lifetimes)]
     #[allow(unreachable_pub)]