+++ tests/expand/not_unpin/tuple_struct.expanded.rs
@@ -2,6 +2,7 @@ use pin_project::pin_project;
 # [pin (__private (! Unpin))]
 struct TupleStruct<T, U>(#[pin] T, U);
 #[allow(box_pointers)]
+#[allow(deprecated)]
 #[allow(explicit_outlives_requirements)]
 #[allow(single_use_lifetimes)]
 #[allow(unreachable_pub)]
@@ -11,6 +12,7 @@ struct TupleStruct<T, U>(#[pin] T, U);
 #[allow(clippy::used_underscore_binding)]
 const _: () = {
     #[allow(box_pointers)]
+    #[allow(deprecated)]
     #[allow(explicit_outlives_requirements)]
     #[allow(single_use_lifetimes)]
     #[allow(unreachable_pub)]
@@ -27,6 +29,7 @@ const _: () = {
     where
         TupleStruct<T, U>: 'pin;
     #[allow(box_pointers)]
+    #[allow(deprecated)]
     #[allow(explicit_outlives_requirements)]
     #[allow(single_use_lifetimes)]
     #[allow(unreachable_pub)]