+++ tests/expand/project_replace/enum.expanded.rs
@@ -10,6 +10,7 @@ enum Enum<T, U> {
     Unit,
 }
 #[allow(box_pointers)]
+#[allow(deprecated)]
 #[allow(explicit_outlives_requirements)]
 #[allow(single_use_lifetimes)]
 #[allow(unreachable_pub)]
@@ -28,6 +29,7 @@ enum EnumProjOwn<T, U> {
     Unit,
 }
 #[allow(box_pointers)]
+#[allow(deprecated)]
 #[allow(explicit_outlives_requirements)]
 #[allow(single_use_lifetimes)]
 #[allow(unreachable_pub)]