+++ CHANGELOG.md
@@ -10,7 +10,9 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
-- Suppress `deprecated` lint in generated code.
+## [1.0.5] - 2021-02-03
+
+- [Suppress `deprecated` lint in generated code.](https://github.com/taiki-e/pin-project/pull/313)
 
 ## [1.0.4] - 2021-01-09
 
@@ -582,7 +584,8 @@ See also [tracking issue for 0.4 release](https://github.com/taiki-e/pin-project
 
 Initial release
 
-[Unreleased]: https://github.com/taiki-e/pin-project/compare/v1.0.4...HEAD
+[Unreleased]: https://github.com/taiki-e/pin-project/compare/v1.0.5...HEAD
+[1.0.5]: https://github.com/taiki-e/pin-project/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/taiki-e/pin-project/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/taiki-e/pin-project/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/taiki-e/pin-project/compare/v1.0.1...v1.0.2